--- a/common/command.cc
+++ b/common/command.cc
@@ -122,6 +122,7 @@ po::options_description CommandHandler::getGeneralOptions()
     general.add_options()("randomize-seed,r", "randomize seed value for random number generator");
     general.add_options()("slack_redist_iter", po::value<int>(), "number of iterations between slack redistribution");
     general.add_options()("cstrweight", po::value<float>(), "placer weighting for relative constraint satisfaction");
+    general.add_options()("starttemp", po::value<float>(), "placer SA start temperature");
     general.add_options()("placer-budgets", "use budget rather than criticality in placer timing weights");
 
     general.add_options()("pack-only", "pack design only without placement or routing");
@@ -187,6 +188,9 @@ void CommandHandler::setupContext(Context *ctx)
 
     if (vm.count("cstrweight")) {
         settings->set("placer1/constraintWeight", vm["cstrweight"].as<float>());
+    }
+    if (vm.count("starttemp")) {
+        settings->set("placer1/startTemp", vm["starttemp"].as<float>());
     }
 
     if (vm.count("placer-budgets")) {