--- a/ice40/examples/floorplan/.gitignore
+++ b/ice40/examples/floorplan/.gitignore
@@ -1,0 +1,4 @@
+*.json
+*.asc
+*.bin
+__pycache__