… CMakeLists.txt
… README.md
… common/placer1.h
… common/placer_heap.cc
… common/placer_heap.h
… common/placer_math.c
… common/placer_math.h
… common/pybindings.cc
… common/pycontainers.h
… common/pywrappers.h
… common/settings.h
… common/timing.cc
… docs/archapi.md
… ecp5/arch.cc
… ecp5/arch.h
… ecp5/arch_pybindings.cc
… ecp5/main.cc
… generic/arch.cc
… generic/arch.h
… gui/fpgaviewwidget.cc
… ice40/arch.cc
… ice40/arch.h
… ice40/arch_pybindings.cc
… ice40/examples/blinky/blinky.pcf
… ice40/examples/blinky/blinky.proj
… ice40/examples/blinky/blinky.sh
… ice40/examples/blinky/blinky.v
… ice40/examples/blinky/blinky.ys
… ice40/examples/blinky/blinky_tb.v
… ice40/examples/floorplan/floorplan.py