Note that is is based on top of #171 (commits up to and including ba6a991) because it uses that parts of that placer for iterative refinement. A decision on merge order remains to be made.

This is a simple analytical placer based on [HeAP](https://janders.eecg.utoronto.ca/pdfs/marcelfpl12.pdf) and [SimPL](http://www.ece.umich.edu/cse/awards/pdfs/iccad10-simpl.pdf). It uses an iterative approach to minimise HPWL, mixed with recursive cut-based spreading followed by an additional strict legalisation pass to  deal with bel validity requirements. In lieu of the iterative refinement described in the paper, after AP finishes regular SA is entered at constant low temperature and diameter (which implements a near-identical algorithm in practice).

On a mid-size iCE40 design (hx8kdemo):
 - master: 85s place time, 7s route time, 33.5MHz Fmax
 - HeAP: 12s place time, 7s route time, 33.4MHz Fmax

On a large ECP5 design (7x picorv32s - 40k LUT4, 26k SLICE) on 85k ECP5:
 - master: 5968s place time, wirelength 199912, 600s route time, 73.99MHz Fmax
 - HeAP: 154s place time, wirlelength 225260, 838s route time, 79.35MHz Fmax