@gsomlo 2019-02-10
The latest/current version of this PR (applied over commit 4a2560b) got my rv64imadc-based blinky demo finished in 17:26 minutes (down from 4:49:42 hours with commit db990d2). I haven't actually tested it by pushing to the FPGA, but will do so early on Monday and report back.