@gsomlo 2019-02-24
eigen3 based version tested on fedora28 and https://github.com/gsomlo/yoloRISC; nextpnr run takes cca 15 minutes, resulting bitstream still working great.