@daveshah1 2019-03-22
This is now (almost) ready for merge from my PoV. The last remaining issue is the failing regression test for #116 - even though HeAP isn't used by default, the general changes to placer1 cause it to now fail. However, it was also failing on master with different seeds so anything that peturbed the PRNG could also cause it to fail. So I think we should neutralise this test until we genuinely fix initial placement  and post-legalisation re-placement.