@eddiehung 2019-03-23
On a mid-size iCE40 design (hx8kdemo) with `--freq 40`:

81a7b5d without `--heap-placer`: 79s place time, 114s route time, 37.47MHz Fmax
81a7b5d with `--heap-placer`: 16s place time, 12s route time, 35.59MHz Fmax

Also with `--seed 0`
81a7b5d without `--heap-placer`: 81s place time, 140s route time, 32.77MHz Fmax
81a7b5d with `--heap-placer`: 14s place time, 9s route time, 38.68MHz Fmax

And with `--seed 1`
81a7b5d without `--heap-placer`: 84s place time, 82s route time, 35.95MHz Fmax
81a7b5d with `--heap-placer`: 17s place time, 10s route time, 34.48MHz Fmax

After this very unrigorous exercise, I think it's safe to say there doesn't seem to be a huge change in final nextpnr-reported Fmax, but it is significant that not just placer time is down, routing time is down too...