@eddiehung 2019-03-23
Okay, looks like we're back to normal now:

52e05f4 without --heap-placer: 118s place time, 10s route time, 31.63MHz Fmax
Also with --seed 0:
52e05f4 without --heap-placer: 115s place time, 12s route time, 31.62MHz Fmax
And with --seed 1:
52e05f4 without --heap-placer: 118s place time, 10s route time, 32.99MHz Fmax