@daveshah1 2019-03-24
Added the `--placer` option and moved stats to one line. Looking at nextpnr-bench now.