@daveshah1 2019-03-24
nextpnr-bench results

arachne-pnr baseline:
```
design01-hx8k average Fmax: 25.98MHz runtime: 69.9s place: 30.6s route: 39.2s
design01-hx8k stddev Fmax: 2.01MHz runtime: 25.4s place: 1.0s route: 24.9s

design02-hx1k average Fmax: 56.13MHz runtime: 4.2s place: 3.1s route: 1.1s
design02-hx1k stddev Fmax: 4.42MHz runtime: 0.4s place: 0.3s route: 0.3s

design02-hx8k average Fmax: 56.32MHz runtime: 9.2s place: 3.7s route: 5.5s
design02-hx8k stddev Fmax: 2.48MHz runtime: 0.4s place: 0.5s route: 0.5s

design03-up5k average Fmax: 6.39MHz runtime: 43.2s place: 21.2s route: 22.0s
design03-up5k stddev Fmax: 0.20MHz runtime: 2.7s place: 1.2s route: 2.3s
```

master:
```
design01-hx8k average Fmax: 35.80MHz runtime: 97.5s place: 89.3s route: 8.2s
design01-hx8k stddev Fmax: 0.82MHz runtime: 2.7s place: 2.3s route: 1.2s

design02-hx1k average Fmax: 71.50MHz runtime: 8.6s place: 7.8s route: 0.8s
design02-hx1k stddev Fmax: 2.84MHz runtime: 0.8s place: 0.7s route: 0.4s

design02-hx8k average Fmax: 72.24MHz runtime: 12.6s place: 11.5s route: 1.1s
design02-hx8k stddev Fmax: 2.28MHz runtime: 0.9s place: 0.9s route: 0.3s

design03-up5k average Fmax: 8.25MHz runtime: 65.7s place: 50.2s route: 15.5s
design03-up5k stddev Fmax: 0.22MHz runtime: 1.9s place: 0.6s route: 1.8s
```

new SA:
```
design01-hx8k average Fmax: 40.48MHz runtime: 142.9s place: 125.0s route: 17.9s
design01-hx8k stddev Fmax: 1.23MHz runtime: 7.1s place: 5.3s route: 7.1s

design02-hx1k average Fmax: 76.48MHz runtime: 7.3s place: 6.6s route: 0.7s
design02-hx1k stddev Fmax: 1.64MHz runtime: 0.5s place: 0.5s route: 0.5s

design02-hx8k average Fmax: 76.69MHz runtime: 8.0s place: 7.0s route: 1.0s
design02-hx8k stddev Fmax: 4.16MHz runtime: 0.4s place: 0.4s route: 0.0s

design03-up5k average Fmax: 8.66MHz runtime: 91.5s place: 66.0s route: 25.5s
design03-up5k stddev Fmax: 0.22MHz runtime: 4.1s place: 1.5s route: 3.7s
```

HeAP:
```
design01-hx8k average Fmax: 40.08MHz runtime: 24.6s place: 15.7s route: 8.9s
design01-hx8k stddev Fmax: 0.71MHz runtime: 2.4s place: 1.3s route: 1.4s

design02-hx1k average Fmax: 72.71MHz runtime: 5.4s place: 4.5s route: 0.9s
design02-hx1k stddev Fmax: 2.14MHz runtime: 1.0s place: 1.2s route: 0.3s

design02-hx8k average Fmax: 75.37MHz runtime: 3.3s place: 2.2s route: 1.1s
design02-hx8k stddev Fmax: 2.07MHz runtime: 0.8s place: 0.9s route: 0.3s

design03-up5k average Fmax: 8.56MHz runtime: 33.5s place: 18.9s route: 14.6s
design03-up5k stddev Fmax: 0.17MHz runtime: 4.0s place: 3.0s route: 2.3s
```

tl;dr the criticality changes, etc, do make the SA option slower for some designs - but with an increase in Fmax on the other hand. I don't know if the existence of the now-much-faster HeAP option for those who care about runtime is enough to justify this. Hopefully in a month or two we can make HeAP the default anyway.