@daveshah1 2019-03-24
nextpnr-bench with new SA and `--placer-budgets`

```
design01-hx8k average Fmax: 36.29MHz runtime: 99.8s place: 92.9s route: 6.9s
design01-hx8k stddev Fmax: 0.78MHz runtime: 4.0s place: 4.3s route: 1.4s

design02-hx1k average Fmax: 74.16MHz runtime: 6.0s place: 5.6s route: 0.4s
design02-hx1k stddev Fmax: 1.75MHz runtime: 0.4s place: 0.5s route: 0.5s

design02-hx8k average Fmax: 72.12MHz runtime: 6.6s place: 6.0s route: 0.6s
design02-hx8k stddev Fmax: 1.83MHz runtime: 0.5s place: 0.4s route: 0.5s

design03-up5k average Fmax: 7.55MHz runtime: 72.5s place: 52.0s route: 20.5s
design03-up5k stddev Fmax: 0.12MHz runtime: 6.0s place: 2.1s route: 5.8s
```