@janrinze 2019-03-24
Is there a specific reason why the speedup on up5k is so much smaller? I expected a bigger difference between master and HeAP.
With SA on hx8k compared to master we get a 36% higher Fmax compared to master.
With SA on up5k the resulting Fmax is only 5% more compared to master.

A possible explanation that i was thinking of is that the delay contributed by the placement is much shorter than the delay of the LUTs and other component. Which in turn would mean that only a small overall reduction in signal propagation can be achieved through better placement on up5k.
The difference with arachne-pnr does however seem to paint a different picture.

The reduced processing time of HeAP is a very welcome result.