@daveshah1 2019-03-24
@janrinze FYI, the design03-up5k benchmark is a totally different design (a NES core) to the design01-hx8k one (picosoc). Indeed you are correct, the design03-up5k test has a high number of logic levels (and is pushing the device to the limits in terms of utilisation), so there is little a placer can do.