@janrinze 2019-03-24
one a side note: does any of the timing rules take into account if there are mixed uses of posedge and negedge of the same clock?