@daveshah1 2019-03-24
Yes, that is considered inside nextpnr's timing analysis (unlike icetime) and Fmax will be calculated correctly where the critical path for a clock is between different edges.